Restructure feedback API to follow admin/user separation pattern
- Move admin-only feedback operations from /api/feedback to /api/admin/feedback
- Keep user feedback submission at /api/feedback (POST only)
- Move individual feedback operations to /api/admin/feedback/[id] (GET, PATCH, DELETE, POST)
- Update frontend components to use new admin endpoints
- Restructure test files to match new API organization
- Maintain comprehensive test coverage with proper endpoint separation
- Follow established pattern used by other admin endpoints in the codebase

This creates clear separation between user-facing and admin-only operations
while maintaining consistent architecture throughout the application.
